refactor: collection card style
