+++ components/collection/CollectionList.vue
