+++ locales/en.json
@@ -1077,9 +1077,7 @@
   },
   "collectionCard": {
     "volume": "Volume",
-    "highestSale": "Highest Sale",
     "items": "Items"
-
   },
   "tabs": {
     "properties": "Properties",
